+++ data/src/main/java/com/imoviedb/app/data/repository/authentication/normaluser/LoginRepositoryImpl.kt
@@ -1,13 +1,18 @@
 package com.imoviedb.app.data.repository.authentication.normaluser
 
-import com.imoviedb.app.data.networking.utils.DispatcherProvider
-import com.imoviedb.app.data.dto.authentication.mapper.AccessTokenValidateMapper
-import com.imoviedb.app.data.dto.authentication.mapper.NewSessionMapper
+import com.imoviedb.app.data.dto.base.ErrorResponseDto
+import com.imoviedb.app.data.dto.authentication.mapper.accesstoken.AccessTokenValidateDtoModelMapper
+import com.imoviedb.app.data.dto.authentication.mapper.accesstoken.AccessTokenValidateErrorModelMapper
+import com.imoviedb.app.data.dto.authentication.mapper.accesstoken.AccessTokenValidateModelEntityMapper
+import com.imoviedb.app.data.dto.authentication.mapper.newsession.NewSessionDtoDomainMapper
+import com.imoviedb.app.data.dto.authentication.mapper.newsession.NewSessionErrorDtoModelMapper
+import com.imoviedb.app.data.dto.authentication.mapper.newsession.NewSessionModelEntityMapper
 import com.imoviedb.app.data.networking.apiservice.AuthenticationService
-import com.imoviedb.app.domain.account.model.AuthenticationBody
+import com.imoviedb.app.data.networking.utils.toErrorModel
 import com.imoviedb.app.data.storage.authentication.UserSessionDAO
 import com.imoviedb.app.data.storage.authentication.UserTokenDAO
-import kotlinx.coroutines.Dispatchers
+import com.imoviedb.app.domain.account.model.AuthenticationBody
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
 import javax.inject.Inject
@@ -16,34 +21,53 @@ class LoginRepositoryImpl @Inject constructor(
     private val userTokenDAO: UserTokenDAO,
     private val authenticationService: AuthenticationService,
     private val userSessionDAO: UserSessionDAO,
-    private val dispatcherProvider: DispatcherProvider,
-    private val accessTokenValidateMapper: AccessTokenValidateMapper,
-    private val sessionMapper: NewSessionMapper
+    private val accessTokenDtoModelMapper: AccessTokenValidateDtoModelMapper,
+    private val accessTokenModelEntityMapper: AccessTokenValidateModelEntityMapper,
+    private val accessTokenErrorModelMapper: AccessTokenValidateErrorModelMapper,
+
+    private val newSessionDtoDomainMapper: NewSessionDtoDomainMapper,
+    private val newSessionErrorDtoModelMapper: NewSessionErrorDtoModelMapper,
+    private val newSessionModelEntityMapper: NewSessionModelEntityMapper
+
+    ,
 
 ) : com.imoviedb.app.domain.authentication.normaluser.repository.LoginRepository {
 
-    override suspend fun validateUserCredential(authenticationBody: AuthenticationBody) = flow {
-        authenticationService.authenticateUserDetails(requestBody = authenticationBody.asMap())
-            .body()?.let {
+    override suspend fun validateUserCredential(authenticationBody: AuthenticationBody,coroutineDispatcher: CoroutineDispatcher) = flow {
+        val response = authenticationService.authenticateUserDetails(requestBody = authenticationBody.asMap())
+        if(response.isSuccessful && response.body() != null){
+            response.body()?.let {
                 it.requestToken?.let { _ ->
-                    val domainModel = accessTokenValidateMapper.mapDtoToModel(it)
-                    userTokenDAO.saveToken(accessTokenValidateMapper.mapModelToEntity(domainModel))
+                    val domainModel = accessTokenDtoModelMapper.map(it)
+                    userTokenDAO.saveToken(accessTokenModelEntityMapper.map(domainModel))
                     emit(domainModel)
                 }
             }
-    }.flowOn(dispatcherProvider.io)
+        }else{
+            val errorModel = response.errorBody()!!.toErrorModel<ErrorResponseDto>()
+            emit(accessTokenErrorModelMapper.map(errorModel))
+        }
+    }.flowOn(coroutineDispatcher)
 
 
-    override suspend fun createNewSessionIDForUser(requestBody: HashMap<String, String>) =
+    override suspend fun createNewSessionIDForUser(requestBody: HashMap<String, String>,coroutineDispatcher: CoroutineDispatcher) =
         flow {
-            authenticationService.createSessionID(requestBody = requestBody).body()?.let {
-                it.sessionId?.let { _ ->
-                    userSessionDAO.removeAllSessions()
-                    val domainModel = sessionMapper.dtoToModel(it)
-                    userSessionDAO.saveSession(sessionMapper.modelToEntity(domainModel))
-                    emit(domainModel)
+            val response = authenticationService.createSessionID(requestBody = requestBody)
+            if(response.isSuccessful && response.body() != null){
+                response.body()?.let {
+                    it.sessionId?.let { _ ->
+                        userSessionDAO.removeAllSessions()
+                        val domainModel = newSessionDtoDomainMapper.map(it)
+                        userSessionDAO.saveSession(newSessionModelEntityMapper.map(domainModel))
+                        emit(domainModel)
+                    }
                 }
+            }else{
+                val errorModel = response.errorBody()!!.toErrorModel<ErrorResponseDto>()
+                emit(newSessionErrorDtoModelMapper.map(errorModel))
             }
-        }.flowOn(Dispatchers.IO)
+
+
+        }.flowOn(coroutineDispatcher)
 
 }
